pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "catan-sdk"
version = "0.1.0"
description = "Catan game engine and bot development toolkit"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-timeout",
    "tqdm",
    "httpx",
]

[tool.hatch.build.targets.wheel]
packages = ["catan"]
